requirements.txt
Django>=5.1,<6.1
djangorestframework>=3.15.2
drf-spectacular>=0.28.0
channels>=4.1.0
channels-redis>=4.2.0
psycopg[binary]>=3.2.3
celery>=5.4.0
redis>=5.2.0
stripe>=11.4.1
gunicorn>=23.0.0
python-dotenv>=1.0.1
pytest>=8.0.0
pytest-django>=4.7.0
